Make FileLogger not overwrite already-written bytes
FileLogger was double-printing newlines and then overwriting the second newline with the first byte of the next line.  By fseeking to the very end and removing the extra newline, we can make the data read by 'tail -f' match the data that eventually gets written to the file.

H/T @flopex for finding the issue.